+++ src/types/librato/IRetrieveMetricResponse.ts
@@ -1,0 +1,4 @@
+export interface IRetrieveMetricResponse {
+  type: 'gauge' | 'composite';
+  name: string;
+}