+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.0.2
+  * Remove IDE directories from package
+  * Make errors from librato service useful
+  * Create metrics from conditions, if they do not exist
+
 ### 1.0.1
   * Add authentication header to librato requests